# gen/http/openapi.yaml
swagger: "2.0"
info:
  title: Messari Metrics
  description: Service providing metrics on single and aggregate assets from Messari Data.
  version: ""
host: localhost:8000
consumes:
- application/json
- application/xml
- application/gob
produces:
- application/json
- application/xml
- application/gob
paths:
  /aggregate:
    get:
      tags:
      - metrics
      summary: aggregate metrics
      operationId: metrics#aggregate
      parameters:
      - name: tags
        in: query
        required: false
        type: string
      - name: sector
        in: query
        required: false
        type: string
      responses:
        "200":
          description: OK response.
          schema:
            type: array
            items:
              $ref: '#/definitions/AssetMetrics'
      schemes:
      - http
  /asset/{slug}:
    get:
      tags:
      - metrics
      summary: asset metrics
      operationId: metrics#asset
      parameters:
      - name: slug
        in: path
        required: true
        type: string
      responses:
        "200":
          description: OK response.
          schema:
            $ref: '#/definitions/MetricsAssetResponseBody'
      schemes:
      - http
definitions:
  AssetMetrics:
    title: AssetMetrics
    type: object
    properties:
      assetSlug:
        type: string
        description: Asset slug
        example: Deleniti omnis.
      chg24hr:
        type: number
        description: Change in price over last 24 hours
        example: 0.22824500320696345
        format: double
      mktcap:
        type: number
        description: Market cap of asset
        example: 0.5357941098735186
        format: double
      price:
        type: number
        description: Current spot price in USD
        example: 0.9974494704262243
        format: double
      vlm24hr:
        type: number
        description: Volume traded over last 24 hours
        example: 0.2827616709063785
        format: double
    example:
      assetSlug: Qui aut impedit ipsa omnis et.
      chg24hr: 0.5116974205645489
      mktcap: 0.0160442318077319
      price: 0.7364241016375906
      vlm24hr: 0.7527376846536737
  AssetMetricsResponseBody:
    title: AssetMetricsResponseBody
    type: object
    properties:
      assetSlug:
        type: string
        description: Asset slug
        example: Distinctio doloribus ab a occaecati.
      chg24hr:
        type: number
        description: Change in price over last 24 hours
        example: 0.31418984381071996
        format: double
      mktcap:
        type: number
        description: Market cap of asset
        example: 0.07264641436963874
        format: double
      price:
        type: number
        description: Current spot price in USD
        example: 0.5287576957572823
        format: double
      vlm24hr:
        type: number
        description: Volume traded over last 24 hours
        example: 0.9377502259503253
        format: double
    example:
      assetSlug: Aut ea.
      chg24hr: 0.5108998417217978
      mktcap: 0.3382291708116491
      price: 0.46584838432716885
      vlm24hr: 0.1330519943929049
  MetricsAssetResponseBody:
    title: MetricsAssetResponseBody
    $ref: '#/definitions/AssetMetricsResponseBody'